requirements.txt
requests==2.32.5
pandas==2.3.3
python-dotenv==1.2.1
pytest==8.4.2
jupyter==1.1.1
matplotlib==3.9.4
numpy==2.0.2